repo-tools: move sql-reports to separate dir
